% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersionShiftFunction.R
\name{dispersionShift}
\alias{dispersionShift}
\title{Alters Mulitvariate Dispersion of Point CLoud within Simplex}
\usage{
dispersionShift(tbl2, dispFact = 1)
}
\arguments{
\item{tbl2}{a closed and imputed data matrix of points cloud}

\item{dispFact}{Factor to increase (dispFact>0) or decrease (dispFact<0) dispersion by}
}
\value{
A list containing:\tabular{ll}{
   \code{combinedOutput} \tab xxx xx xx  \cr
   \tab \cr
   \code{dispAdjustedData} \tab xxx xxx \cr
    \tab \cr
   \code{dispDistr.shifted} \tab xxx \cr
    \tab \cr
   \code{dispDistr.adj} \tab xxx \cr
}
}
\description{
Alters Mulitvariate Dispersion of Point CLoud within Simplex
}
