% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearShiftFunction.R
\name{eigenDecomp.CLR}
\alias{eigenDecomp.CLR}
\title{Eigen decompositon of Composiitonal Dataset}
\usage{
eigenDecomp.CLR(compMat)
}
\arguments{
\item{compMat}{a compositional data matrix/frame}
}
\value{
A list containing:\tabular{ll}{
   \code{clrmat} \tab clr transformed data  \cr
   \tab \cr
   \code{Coords} \tab Eigendecomposition coordinates \cr
    \tab \cr
   \code{eigenVector} \tab eigenvectors \cr
    \tab \cr
   \code{eigenValues} \tab eigenvalues \cr
    \tab \cr
   \code{percentExplained} \tab percent of variation explained by each component \cr
}
}
\description{
Performs eigen decompositon on compositional dataset after imputing zeroes and clr transform
}
