% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnGraph_Functions.R
\name{knn_graph}
\alias{knn_graph}
\title{Computes a K-Farthest / Nearest Neighbor Graph}
\usage{
knn_graph(adj_mat, K, plot_TrueFalse = FALSE, sim_ = T)
}
\arguments{
\item{adj_mat}{pxp Adjacency Matrix}

\item{K}{K farthest neighbors}

\item{plot_TrueFalse}{should results be plotted}

\item{sim_}{returns a K nearest neighbor graph}
}
\value{
A list containing:\tabular{ll}{
   \code{Graph} \tab the igraph graph object  \cr
   \tab \cr
   \code{adj} \tab The KFN or KNN weighted adjacency matrix
}
}
\description{
Compute a K-Nearest/Farthest Neighbor (KFN / KNN) graph from a weighted adjacency matrix
}
\examples{
\dontrun{
## compute pairwise distance
comp = simAddLogNormal(n1 = 30)[,-1]
dd = dist(compositions::clr(comp))
## Compute 5-KNN Graph
g = knn_graph(as.matrix(dd),K = 5,sim_ = T,plot_TrueFalse = T)
}
}
