% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnGraph_Functions.R
\name{knnADJtoSYM}
\alias{knnADJtoSYM}
\title{Make Symetric}
\usage{
knnADJtoSYM(knnADJ_MAT)
}
\arguments{
\item{knnADJ_MAT}{knn / kfn dj. matrix}
}
\value{
a symetric knn/kfn weighted adj matrix
}
\description{
Converts a KNN/KFN adjaceny matrix to a symetric matrix
}
