% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearShiftFunction.R
\name{linearShift}
\alias{linearShift}
\title{Linearly Shift a Point Cloud in the Simplex}
\usage{
linearShift(
  tbl2,
  pertubationDirection = NULL,
  directionType = "centroid",
  evComp = 1,
  a = 1,
  alpha_n = 1
)
}
\arguments{
\item{tbl2}{closed and imputed data matrix to be shifted within the simplex}

\item{pertubationDirection}{A user defined direction to shift point cloud; default = NULL}

\item{directionType}{The direction type of a standar shift in the simplex; either ("Centroid", "Eigen")}

\item{evComp}{if the directionType=="eigen" then which component to use for the linear shift; deafult = 1 for the first eigenvector}

\item{a}{A vector of \eqn{\alpha} power coeeficients to scale the linear shift; determines the distance of the shift}

\item{alpha_n}{Return the point cloud shift corresponding to a specific\eqn{\alpha} coefficient}
}
\value{
A list containing:\tabular{ll}{
   \code{combinedOutput} \tab xxx  \cr
   \tab \cr
   \code{linAdjustedData} \tab xxx \cr
    \tab \cr
   \code{allShiftData} \tab xxx \cr
    \tab \cr
   \code{shiftVec} \tab xxx \cr
    \tab \cr
   \code{meanDelta} \tab xxx \cr
}
}
\description{
Linearly shifts points cloud in the simplex along the directions of the centroid, most variation (eigenvector-1), or a user specified direction
}
