% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simScenarios.R
\name{sampleAddLogisticT}
\alias{sampleAddLogisticT}
\title{Simulate Data additive logisitcs T}
\usage{
sampleAddLogisticT(
  n,
  df_,
  dms,
  sigmaScale = 1,
  sampleName = "S1",
  seed = 8272008
)
}
\arguments{
\item{n}{number of samples to return}

\item{df_}{degrees of freedom}

\item{dms}{number of dimensions}

\item{sigmaScale}{scale covariance matrix}

\item{sampleName}{class name}

\item{seed}{random seed}
}
\value{
a data.frame with simulated data where the first column in the class name
}
\description{
Simulate Data additive logisitcs T
}
\examples{
\dontrun{
# simulate 30 samples from the Dirichlet ditriubtion with default parms
sampleAddLogisticT(n1 = 30)
}
}
