% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simScenarios.R
\name{simAddLogNormal}
\alias{simAddLogNormal}
\title{Simulate Data from Additive Logistics Function}
\usage{
simAddLogNormal(
  n1 = 30,
  dms_ = 75,
  className = "S1",
  seed = 8272008,
  varParm = 2,
  meanVec = 0
)
}
\arguments{
\item{n1}{number of samples to return}

\item{dms_}{number of dimesions to simulate}

\item{className}{name of class from simulated data}

\item{seed}{random seed for simulation}

\item{varParm}{Controls the variance of simulated distribution}

\item{meanVec}{Sets the value of the mean vector for the multivariate distribution}
}
\value{
a data.frame with simulated data where the first column in the class name
}
\description{
Simulates data from the additive logistics normal distribution
}
\examples{
\dontrun{
# simulate 30 samples from the Dirichlet ditriubtion with default parms
simAddLogNormal(n1 = 30)
}
}
