% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCompData4.R
\name{simCompData}
\alias{simCompData}
\title{Simulate / Augment Simplex Bound Data from Emipircal Distribution}
\usage{
simCompData(
  comp,
  labels = NULL,
  k = round(sqrt(nrow(comp))),
  modthres = 0.4,
  ndirections = 1000,
  fct = 1.5,
  alFactor = 1.5,
  ptPerSample = 3
)
}
\arguments{
\item{comp}{empirical composition}

\item{labels}{class labels ( can be extracted later if needed)}

\item{k}{K nearest neighbors for topology matching}

\item{modthres}{modulairty threshold for simulating using community structure}

\item{ndirections}{number of random directions to sample from}

\item{fct}{scale factor for dispersion distance matching}

\item{alFactor}{alpha factor dirichlet concentration parameter vetcor}

\item{ptPerSample}{numer of points to agumnet per sample}
}
\value{
A list containing:\tabular{ll}{
   \code{simData} \tab simulated/augmented data  \cr
   \tab \cr
   \code{alphaData} \tab detailed metrics on distances and aplha power coefficents used to shift points \cr
    \tab \cr
   \code{ClassLabels} \tab Class labels of simulated/aug. data matching the class of origin
}
}
\description{
Simulate / Augment Simplex Bound Data from Emipircal Distribution
}
\examples{
\dontrun{
add example here
}
}
