% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearShiftFunction.R
\name{vecDist}
\alias{vecDist}
\title{Distance Between Compositional Vectors}
\usage{
vecDist(v1, v2)
}
\arguments{
\item{v1}{closed compositional vector 1}

\item{v2}{closed compositional vector 2}
}
\value{
the distance between vecotrs using all pariwise ratios
}
\description{
Compute the distance between vectors using an all pairwise log ratio approach
}
