% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearShiftFunction.R
\name{vecNorm}
\alias{vecNorm}
\title{Compute the Norm of a Closed Compositional Vector}
\usage{
vecNorm(namedVec)
}
\arguments{
\item{namedVec}{a closed compositional vector}
}
\value{
the vector norm
}
\description{
Compute the vector norm of a compositional vector in the simplex using Aitchinson's geometry
}
